chore(act-diagram): modularize tests, add to CI publish matrix, fix CodeQL
- Split evaluate.spec.ts (2275 lines) into 4 focused modules
- Split layout.spec.ts (2082 lines) into 4 focused modules
- 14 test files, 305 tests, 100% coverage maintained
- Add act-diagram to CI/CD publish matrix in ci-cd.yml
- Add CodeQL suppression comment for intentional new Function() usage
- Fix unused imports in test files
