+++ .github/workflows/ci-cd.yml
@@ -59,7 +59,7 @@ jobs:
     strategy:
       max-parallel: 1 # prevents concurrent pull/push conflicts on main branch
       matrix:
-        lib: [act-patch, act, act-sse, act-pg]
+        lib: [act-patch, act, act-sse, act-pg, act-diagram]
 
     steps:
       - uses: actions/checkout@v4
