--- libs/act-diagram/test/evaluate.spec.ts
@@ -1,2275 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { buildModel } from "../src/client/lib/build-model.js";
-import { extractModel } from "../src/client/lib/evaluate.js";
-import type { FileTab } from "../src/client/types/file-tab.js";
-
-describe("extractModel", () => {
-  it("extracts a simple state with actions and events", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/counter.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-
-export const Counter = state({ Counter: z.object({ count: z.number() }) })
-  .init(() => ({ count: 0 }))
-  .emits({ Incremented: z.object({ amount: z.number() }) })
-  .patch({ Incremented: ({ data }, state) => ({ count: state.count + data.amount }) })
-  .on({ increment: z.object({ by: z.number() }) })
-    .emit("Incremented")
-  .build();
-`,
-      },
-    ];
-
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.states).toHaveLength(1);
-    expect(model.states[0].name).toBe("Counter");
-    expect(model.states[0].events).toHaveLength(1);
-    expect(model.states[0].events[0].name).toBe("Incremented");
-    expect(model.states[0].events[0].hasCustomPatch).toBe(true);
-    expect(model.states[0].actions).toHaveLength(1);
-    expect(model.states[0].actions[0].name).toBe("increment");
-    expect(model.states[0].actions[0].emits).toContain("Incremented");
-  });
-
-  it("extracts slices with reactions and projections", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/states.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-
-export const Ticket = state({ Ticket: z.object({ status: z.string() }) })
-  .init(() => ({ status: "new" }))
-  .emits({
-    TicketOpened: z.object({ title: z.string() }),
-    TicketAssigned: z.object({ to: z.string() }),
-  })
-  .on({ OpenTicket: z.object({ title: z.string() }) })
-    .emit("TicketOpened")
-  .on({ AssignTicket: z.object({ to: z.string() }) })
-    .emit("TicketAssigned")
-  .build();
-`,
-      },
-      {
-        path: "src/proj.ts",
-        content: `
-import { projection } from "@rotorsoft/act";
-import { z } from "zod";
-
-export const TicketProj = projection("tickets")
-  .on({ TicketOpened: z.object({ title: z.string() }) })
-    .do()
-  .build();
-`,
-      },
-      {
-        path: "src/slices.ts",
-        content: `
-import { slice } from "@rotorsoft/act";
-import { Ticket } from "./states.js";
-import { TicketProj } from "./proj.js";
-
-export const TicketSlice = slice()
-  .withState(Ticket)
-  .withProjection(TicketProj)
-  .on("TicketOpened")
-    .do(async function autoAssign(event, _stream, app) {
-      await app.do("AssignTicket", event.stream, { to: "agent" }, event);
-    })
-    .to(() => "default")
-  .build();
-`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { TicketSlice } from "./slices.js";
-import { TicketProj } from "./proj.js";
-
-export const app = act()
-  .withSlice(TicketSlice)
-  .withProjection(TicketProj)
-  .on("TicketOpened")
-    .do(async function logOpen() {})
-    .void()
-  .build();
-`,
-      },
-    ];
-
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.slices).toHaveLength(2);
-    const ticketSlice = model.slices.find((s) => s.name === "TicketSlice")!;
-    expect(ticketSlice.reactions).toHaveLength(1);
-    expect(ticketSlice.reactions[0].dispatches).toContain("AssignTicket");
-    expect(ticketSlice.projections).toContain("tickets");
-    const globalSlice = model.slices.find((s) => s.name === "global")!;
-    expect(globalSlice).toBeDefined();
-    expect(model.entries).toHaveLength(1);
-    expect(model.entries[0].path).toBe("src/app.ts");
-    expect(model.entries[0].projections).toHaveLength(1);
-    expect(model.reactions).toHaveLength(1);
-    expect(model.reactions[0].isVoid).toBe(true);
-    expect(model.orchestrator).toBeDefined();
-  });
-
-  it("extracts standalone projections", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/proj.ts",
-        content: `
-import { projection } from "@rotorsoft/act";
-import { z } from "zod";
-
-export const TicketProjection = projection("tickets")
-  .on({ TicketOpened: z.object({ title: z.string() }) })
-    .do(async ({ stream, data }) => { console.log(stream, data); })
-  .build();
-`,
-      },
-    ];
-
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.projections).toHaveLength(1);
-    expect(model.projections[0].name).toBe("tickets");
-    expect(model.projections[0].handles).toContain("TicketOpened");
-  });
-
-  it("returns empty model for empty files", () => {
-    const { model } = extractModel([]);
-    expect(model.states).toHaveLength(0);
-    expect(model.slices).toHaveLength(0);
-    expect(model.entries).toHaveLength(0);
-  });
-
-  it("handles .tsx and .d.ts and node_modules files (skips them)", () => {
-    const files: FileTab[] = [
-      { path: "src/App.tsx", content: `export const x = 1;` },
-      { path: "src/types.d.ts", content: `declare const y: number;` },
-      { path: "node_modules/foo/index.ts", content: `export const z = 1;` },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(0);
-  });
-
-  it("handles files with syntax errors gracefully", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/broken.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-
-export const Widget = state({ Widget: z.object({ value: z.string() }) })
-  .init(() => ({{ value: "" }}))
-  .emits({ ValueSet: z.object({ value: z.string() }) })
-  .on({ SetValue: z.object({ value: z.string() }) })
-    .emit("ValueSet")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // Broken file produces no states (no regex fallback), but doesn't crash
-    expect(model.states).toHaveLength(0);
-  });
-
-  it("handles import.meta replacement", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-const url = import.meta.url;
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x: "" }))
-  .emits({ Done: z.object({}) })
-  .on({ doIt: z.object({}) }).emit("Done")
-  .build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("resolves relative imports with parent traversal", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/shared/types.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const Shared = state({ Shared: z.object({ v: z.number() }) })
-  .init(() => ({ v: 0 }))
-  .emits({ SharedEvt: z.object({}) })
-  .on({ doShared: z.object({}) }).emit("SharedEvt")
-  .build();
-`,
-      },
-      {
-        path: "src/features/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { Shared } from "../shared/types.js";
-export const app = act().withState(Shared).build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.entries).toHaveLength(1);
-  });
-
-  it("resolves relative /index imports", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/domain/index.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const D = state({ D: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { D } from "./domain.js";
-export const app = act().withState(D).build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.entries).toHaveLength(1);
-  });
-
-  it("resolves scoped package imports to workspace files", () => {
-    const files: FileTab[] = [
-      {
-        path: "packages/utils/src/index.ts",
-        content: `export const helper = "hello";`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { helper } from "@myorg/utils";
-const x = helper;
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x }))
-  .emits({ Done: z.object({}) })
-  .on({ doIt: z.object({}) }).emit("Done")
-  .build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("resolves bare unknown module to unknownModuleProxy", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import pkg from "totally-unknown-package";
-const x = pkg.something();
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("strips runtime calls (main, run, start, bootstrap)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state, act } from "@rotorsoft/act";
-import { z } from "zod";
-const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-const app = act().withState(S).build();
-async function main() { await app.do("a", { stream: "s", actor: { id: "1", name: "u" } }, {}); }
-main().catch(console.error);
-void run();
-await start();
-bootstrap();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states.length).toBeGreaterThanOrEqual(1);
-  });
-
-  it("handles __dirname/__filename const declarations", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-const __dirname = "/some/path";
-const __filename = "/some/path/file.ts";
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x: "" }))
-  .emits({ Done: z.object({}) })
-  .on({ doIt: z.object({}) }).emit("Done")
-  .build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("creates default entry when no act() but states exist", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/states.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({ x: z.number() }) })
-  .init(() => ({ x: 0 }))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.entries).toHaveLength(1);
-    expect(model.entries[0].path).toBe("app");
-  });
-
-  it("handles store() and dispose() imports", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state, act, store, dispose } from "@rotorsoft/act";
-import { z } from "zod";
-store();
-const d = dispose();
-const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-const app = act().withState(S).build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("handles node:crypto import", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { randomUUID } from "node:crypto";
-const id = randomUUID();
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x: id }))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("resolves root-level files (no directory) with relative imports", () => {
-    const files: FileTab[] = [
-      {
-        path: "types.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const T = state({ T: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-`,
-      },
-      {
-        path: "app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { T } from "./types.js";
-export const app = act().withState(T).build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.entries).toHaveLength(1);
-  });
-
-  // --- Regex fallback branches ---
-
-  it("broken file with only states produces nothing", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ Created: z.object({}), Updated: z.object({}) })
-  .patch({ Created: (e, s) => s, Updated: (e, s) => s })
-  .on({ create: z.object({}) }).emit("Created")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(0);
-  });
-
-  it("regex fallback extracts emit from arrow function handler with single-quoted event", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ MyEvent: z.object({}) })
-  .on({ doIt: z.object({}) })
-    .emit((action) => ['MyEvent', { data: action }])
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    const st = model.states.find((s) => s.name === "S");
-    if (st) {
-      const action = st.actions.find((a) => a.name === "doIt");
-      if (action) expect(action.emits).toContain("MyEvent");
-    }
-  });
-
-  it("broken file with slice produces error placeholder", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { slice, state, projection } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-const S = state({ S: z.object({}) }).init(() => ({})).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-const P = projection("proj").on({ E: z.object({}) }).do().build();
-const MySlice = slice()
-  .withState(S)
-  .withProjection(P)
-  .on("E").do(async function handler() {}).void()
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.slices).toHaveLength(1);
-    expect(model.slices[0].error).toBeDefined();
-  });
-
-  it("regex fallback extracts given invariants", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ Done: z.object({}) })
-  .on({ doIt: z.object({}) })
-    .given([{ description: "must be valid", valid: () => true }])
-    .emit("Done")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    const st = model.states.find((s) => s.name === "S");
-    if (st) {
-      const action = st.actions.find((a) => a.name === "doIt");
-      if (action) expect(action.invariants).toContain("must be valid");
-    }
-  });
-
-  it("broken file does not produce projections", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { projection } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-const P = projection("myProj")
-  .on({ A: z.object({}) }).do()
-  .on({ B: z.object({}) }).do()
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.projections).toHaveLength(0);
-  });
-
-  // --- Model building branch coverage ---
-
-  it("handles state without events or actions (null guards)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({})
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(1);
-    expect(model.states[0].events).toHaveLength(0);
-    expect(model.states[0].actions).toHaveLength(0);
-  });
-
-  it("handles state with event that has no custom patch (hasCustomPatch false)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({ v: z.string() }) })
-  .init(() => ({ v: "" }))
-  .emits({ Done: z.object({ v: z.string() }) })
-  .on({ doIt: z.object({ v: z.string() }) }).emit("Done")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states[0].events[0].hasCustomPatch).toBe(false);
-  });
-
-  it("handles action with invariants that have no description", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) })
-    .given([{ valid: () => true }])
-    .emit("E")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // invariant without description → empty string
-    expect(model.states[0].actions[0].invariants).toContain("");
-  });
-
-  it("returns error when execute produces error", () => {
-    // Force an error by providing code that causes the outer try to fail
-    // This exercises line 414: if (error) return { model, error }
-    // Actually, the outer catch is very hard to trigger - errors from eval
-    // are caught by the inner try. The error path works when execute()
-    // returns an error string.
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({})).emits({}).build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // Just verify it doesn't crash
-    expect(model).toBeDefined();
-  });
-
-  it("relative import to non-existent file returns empty object", () => {
-    // Exercises line 106: ?? {} when relative import doesn't resolve
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { missing } from "./nonexistent.js";
-export const S = state({ S: z.object({}) })
-  .init(() => ({})).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("scoped import resolution falls through all patterns to unknownModuleProxy", () => {
-    // Exercises lines 114-116: all ?? branches in scoped package resolution
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { util } from "@someorg/nonexistent";
-export const S = state({ S: z.object({}) })
-  .init(() => ({})).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("broken file with only states and backtick emits produces nothing", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content:
-          'import { state } from "@rotorsoft/act";\nimport { z } from "zod";\nconst x = ({{ broken }});\nexport const S = state({ S: z.object({}) })\n  .init(() => ({}))\n  .emits({ MyEvt: z.object({}) })\n  .on({ doIt: z.object({}) })\n    .emit((a) => [`MyEvt`, {}])\n  .build();\n',
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(0);
-  });
-
-  it("broken file with slice produces error placeholder (variable handler)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { slice } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-const MySlice = slice()
-  .on("SomeEvent").do(myHandler)
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.slices).toHaveLength(1);
-    expect(model.slices[0].error).toBeDefined();
-  });
-
-  it("regex fallback: slice reaction with anonymous handler", () => {
-    // Exercises line 353: `on ${odm[1]}` fallback when no function name
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { slice } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-const MySlice = slice()
-  .on("SomeEvent").do(async () => {})
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // The regex won't capture anonymous arrow, so no reactions extracted
-    expect(model).toBeDefined();
-  });
-
-  it("handles act with scoped package imports (not @rotorsoft)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state, act } from "@rotorsoft/act";
-import { z } from "zod";
-import { something } from "@myorg/utils";
-const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-const app = act().withState(S).build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("scoped import resolves to pkg/src/index pattern", () => {
-    const files: FileTab[] = [
-      {
-        path: "utils/src/index.ts",
-        content: `export const helper = "val";`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { helper } from "@org/utils";
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x: helper }))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("scoped import resolves to pkg/index pattern", () => {
-    const files: FileTab[] = [
-      {
-        path: "utils/index.ts",
-        content: `export const helper = "val";`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { helper } from "@org/utils";
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x: helper }))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("scoped import resolves to bare pkg pattern", () => {
-    const files: FileTab[] = [
-      {
-        path: "utils.ts",
-        content: `export const helper = "val";`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { helper } from "@org/utils";
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x: helper }))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("relative import resolves to /index fallback", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/domain/index.ts",
-        content: `export const val = 42;`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { val } from "./domain.js";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("slice varName tagging skips acts from other files", () => {
-    // act() in app.ts has .withSlice(MySlice), but slice is built in slices.ts
-    // The tagging loop should skip acts whose _sourceFile !== current file
-    const files: FileTab[] = [
-      {
-        path: "src/states.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-      {
-        path: "src/slices.ts",
-        content: `
-import { slice } from "@rotorsoft/act";
-import { S } from "./states.js";
-export const MySlice = slice().withState(S).build();
-`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { MySlice } from "./slices.js";
-export const app = act().withSlice(MySlice).build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.slices).toHaveLength(1);
-    expect(model.slices[0].name).toBe("MySlice");
-  });
-
-  it("broken file produces error placeholder for slices defined in it", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/b.ts",
-        content: `
-import { state, slice } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ Done: z.object({}) })
-  .on({ doIt: z.object({}) }).emit("Done")
-  .build();
-export const MySlice = slice()
-  .withState(S)
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(0);
-    expect(model.slices).toHaveLength(1);
-    expect(model.slices[0].error).toBeDefined();
-  });
-
-  it("fixes up reaction handler names from source when mock yields 'on EventName'", () => {
-    // When the handler module can't be resolved (circular dep or missing),
-    // handler.name isn't a real string and falls back to "on EventName".
-    // The fixup scans the source for .on("Event").do(module.handler) to recover.
-    const files: FileTab[] = [
-      {
-        path: "domain/src/index.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ Created: z.object({}) })
-  .on({ create: z.object({}) }).emit("Created")
-  .build();
-`,
-      },
-      {
-        path: "app/src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { S } from "@org/domain";
-import { onCreated } from "unresolvable-external-package";
-export const app = act()
-  .withState(S)
-  .on("Created").do(onCreated).to("items")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.reactions).toHaveLength(1);
-    // handler.name on proxy isn't a string → falls back to "on Created"
-    // fixupReactions recovers "onCreated" from source
-    expect(model.reactions[0].handlerName).toBe("onCreated");
-  });
-
-  it("broken file does not produce projections (no fallback scan)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/proj.ts",
-        content: `
-import { projection } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken syntax }});
-export const P = projection("uncaptured")
-  .on({ A: z.object({}) }).do()
-  .on({ B: z.object({}) }).do()
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.projections).toHaveLength(0);
-  });
-
-  it("projection fallback scan skips already-captured projections", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/proj.ts",
-        content: `
-import { projection } from "@rotorsoft/act";
-import { z } from "zod";
-export const P = projection("items")
-  .on({ ItemCreated: z.object({}) }).do()
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // Mock eval captures it, fallback scan should not duplicate
-    expect(model.projections).toHaveLength(1);
-  });
-
-  it("broken projection file produces no projections in act entry", () => {
-    // When the projection file is broken, the act entry has no projections
-    const files: FileTab[] = [
-      {
-        path: "src/proj.ts",
-        content: `
-import { projection } from "@rotorsoft/act";
-import { z } from "zod";
-const x = ({{ broken }});
-export const GameProjection = projection("games")
-  .on({ GameCreated: z.object({}) }).do()
-  .build();
-`,
-      },
-      {
-        path: "src/states.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const Game = state({ Game: z.object({}) })
-  .init(() => ({}))
-  .emits({ GameCreated: z.object({}) })
-  .on({ CreateGame: z.object({}) }).emit("GameCreated")
-  .build();
-`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { Game } from "./states.js";
-import { GameProjection } from "./proj.js";
-export const app = act()
-  .withState(Game)
-  .withProjection(GameProjection)
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // Broken projection file produces no projections
-    expect(model.entries).toHaveLength(1);
-    expect(model.entries[0].projections).toHaveLength(0);
-  });
-
-  it("null guards handle act with undefined projections/states/slices/reactions", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-export const app = act().build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.entries).toHaveLength(1);
-  });
-
-  it("skips test and spec files during extraction", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-      {
-        path: "src/app.spec.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const TestState = state({ TestState: z.object({}) })
-  .init(() => ({}))
-  .emits({ TestEvt: z.object({}) })
-  .on({ test: z.object({}) }).emit("TestEvt").build();
-`,
-      },
-      {
-        path: "src/__tests__/helper.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const Helper = state({ Helper: z.object({}) })
-  .init(() => ({}))
-  .emits({ H: z.object({}) })
-  .on({ h: z.object({}) }).emit("H").build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // Only the non-test state should be extracted
-    expect(model.states).toHaveLength(1);
-    expect(model.states[0].name).toBe("S");
-  });
-
-  it("handles act() with anonymous inline reaction", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state, act } from "@rotorsoft/act";
-import { z } from "zod";
-const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-const app = act()
-  .withState(S)
-  .on("E")
-    .do(async (_ev, _s, app) => { await app.do("a", "s", {}); })
-    .to(() => "t")
-  .build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.reactions).toHaveLength(1);
-    expect(model.reactions[0].isVoid).toBe(false);
-  });
-
-  it("fixupReactions joins all files when no sourceFile provided (line 66)", () => {
-    // Directly test buildModel: slice reactions with fallback "on EventName"
-    // handler names get fixed up by scanning all file contents when no sourceFile
-    //
-    const files: FileTab[] = [
-      {
-        path: "src/slices.ts",
-        content: `.on("Created").do(onCreated).to(() => "t")`,
-      },
-    ];
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "MySlice",
-          states: [],
-          projections: [],
-          reactions: [
-            {
-              event: "Created",
-              handlerName: "on Created",
-              dispatches: [],
-              isVoid: false,
-            },
-          ],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, files, new Map());
-    const sl = model.slices.find((s) => s.name === "MySlice");
-    expect(sl).toBeDefined();
-    // The "on Created" handler name should be fixed up to "onCreated"
-    expect(sl!.reactions[0].handlerName).toBe("onCreated");
-  });
-
-  it("act builder states not already in stateByRef (lines 134-136)", () => {
-    // Use buildModel directly: state only in act.states, not in result.states
-    //
-    const actOnlyState = {
-      _tag: "State",
-      name: "ActOnlyState",
-      events: { Evt1: {} },
-      actions: { doIt: {} },
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [], // state NOT in rawStates
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [actOnlyState], // state only here
-          slices: [],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const st = model.states.find((s) => s.name === "ActOnlyState");
-    expect(st).toBeDefined();
-    expect(st!.events).toHaveLength(1);
-    expect(st!.events[0].name).toBe("Evt1");
-  });
-
-  it("slice with null state reference produces Missing state reference (lines 166-167)", () => {
-    // Directly test buildModel with a raw slice that has null in states array
-    //
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "BrokenSlice",
-          states: [null, undefined],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "BrokenSlice");
-    expect(sl).toBeDefined();
-    expect(sl!.error).toContain("broken import");
-  });
-
-  it("slice state not in stateByRef, tries addState in-place success (lines 173-180)", () => {
-    // Slice references a state that wasn't built in step 1
-    //
-    const inlineState = {
-      _tag: "State",
-      name: "InlineSliceState",
-      events: { Evt1: {} },
-      actions: { doIt: {} },
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [], // state NOT pre-built
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "TestSlice",
-          states: [inlineState],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "TestSlice");
-    expect(sl).toBeDefined();
-    expect(sl!.states.length).toBe(1);
-    // The state should have been added to model.states
-    const st = model.states.find((s) => s.name === "InlineSliceState");
-    expect(st).toBeDefined();
-  });
-
-  it("slice state not in stateByRef, addState fails (lines 181-182)", () => {
-    // Slice references a state not in stateByRef, but addState throws
-    //
-    const corruptState = {
-      name: "CorruptState",
-      get events(): any {
-        throw new Error("corrupt events");
-      },
-      actions: {},
-    };
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "SliceWithCorrupt",
-          states: [corruptState],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "SliceWithCorrupt");
-    expect(sl).toBeDefined();
-    expect(sl!.error).toContain("corrupt");
-    expect(sl!.error).toContain("corrupt events");
-  });
-
-  it("slice processing catch block when entire slice build throws (lines 204-205)", () => {
-    //
-    // Create a slice object whose states getter throws
-    const throwingSlice = {
-      _tag: "Slice",
-      _varName: "ThrowSlice",
-      get states(): any {
-        throw new Error("boom in states getter");
-      },
-      projections: [],
-      reactions: [],
-    };
-    const result = {
-      states: [],
-      slices: [throwingSlice],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "ThrowSlice");
-    expect(sl).toBeDefined();
-    expect(sl!.error).toContain("boom in states getter");
-    expect(sl!.states).toHaveLength(0);
-  });
-
-  it("global error return when result.error exists and no states/slices (line 307)", () => {
-    //
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: "Global extraction failed",
-      fileErrors: new Map<string, string>(),
-    };
-    const { model, error } = buildModel(result, [], new Map());
-    expect(error).toBe("Global extraction failed");
-    expect(model.states).toHaveLength(0);
-    expect(model.slices).toHaveLength(0);
-  });
-
-  it("buildState returns error when event has undefined schema (line 31)", () => {
-    const stateWithUndefinedEvent = {
-      _tag: "State",
-      name: "BadState",
-      events: { Evt1: undefined },
-      actions: {},
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [stateWithUndefinedEvent],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    // State with undefined event schema should not appear in model.states
-    expect(model.states.find((s) => s.name === "BadState")).toBeUndefined();
-  });
-
-  it("buildState returns error when action has undefined schema (line 37)", () => {
-    const stateWithUndefinedAction = {
-      _tag: "State",
-      name: "BadActionState",
-      events: { Evt1: {} },
-      actions: { doIt: undefined },
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [stateWithUndefinedAction],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(
-      model.states.find((s) => s.name === "BadActionState")
-    ).toBeUndefined();
-  });
-
-  it("step 1 catch block for corrupted state (line 144)", () => {
-    const corruptState = {
-      _tag: "State",
-      get name(): string {
-        throw new Error("corrupt name getter");
-      },
-      events: {},
-      actions: {},
-    };
-    const result = {
-      states: [corruptState],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    // Should not crash — corrupt state is skipped
-    expect(model).toBeDefined();
-  });
-
-  it("step 1 catch block for act-builder state (line 156)", () => {
-    const corruptState = {
-      _tag: "State",
-      get name(): string {
-        throw new Error("corrupt act state");
-      },
-      events: {},
-      actions: {},
-    };
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [corruptState],
-          slices: [],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model).toBeDefined();
-  });
-
-  it("slice state already built with error in step 1, referenced from slice (line 190)", () => {
-    // State is in rawStates AND in a slice's states array — same object reference
-    // Step 1 builds it and gets an error, then the slice loop finds it via stateByRef
-    const badState = {
-      _tag: "State",
-      name: "BadState",
-      events: { Evt1: undefined }, // undefined schema → buildState returns error
-      actions: {},
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [badState], // step 1 builds this and stores {error} in stateByRef
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "SliceRefBadState",
-          states: [badState], // same object reference — found in stateByRef with error
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "SliceRefBadState");
-    expect(sl).toBeDefined();
-    expect(sl!.error).toContain("undefined schema");
-    expect(sl!.states).toHaveLength(0);
-  });
-
-  it("slice state buildState returns error (line 190)", () => {
-    // State with undefined event schema — buildState returns {error}
-    // when encountered inside a slice that wasn't pre-built
-    const stateWithBadEvent = {
-      _tag: "State",
-      name: "SliceBadEvtState",
-      events: { BadEvt: undefined },
-      actions: {},
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [], // not pre-built
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "SliceWithBadState",
-          states: [stateWithBadEvent],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "SliceWithBadState");
-    expect(sl).toBeDefined();
-    expect(sl!.error).toContain("undefined schema");
-  });
-
-  it("duplicate state in rawStates triggers stateByRef.has continue (line 140)", () => {
-    const stateObj = {
-      _tag: "State",
-      name: "DupState",
-      events: { E: {} },
-      actions: { a: {} },
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [stateObj, stateObj], // same reference twice
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    // Should only produce one state
-    expect(model.states.filter((s) => s.name === "DupState")).toHaveLength(1);
-  });
-
-  it("fixupReactions returns early when sourceFile not in files (line 84)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          _sourceFile: "src/nonexistent.ts",
-          states: [],
-          slices: [],
-          projections: [],
-          reactions: [
-            {
-              event: "Created",
-              handlerName: "on Created",
-              dispatches: [],
-              isVoid: false,
-            },
-          ],
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    // Pass no matching files — fixupReactions should return early
-    const { model } = buildModel(result, [], new Map());
-    // Handler name stays unfixed since source file wasn't found
-    expect(model.reactions[0].handlerName).toBe("on Created");
-  });
-
-  it("act builder skips duplicate state via stateByRef.has (step 1 + act states)", () => {
-    const stateObj = {
-      _tag: "State",
-      name: "SharedState",
-      events: { E: {} },
-      actions: { a: {} },
-      given: {},
-      patches: new Map(),
-    };
-    const result = {
-      states: [stateObj], // pre-built in step 1
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [stateObj], // same reference — already in stateByRef
-          slices: [],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model.states.filter((s) => s.name === "SharedState")).toHaveLength(
-      1
-    );
-  });
-
-  it("act._sourceFile skip in slice varName tagging (line 189)", () => {
-    // Two files: one builds an act, another has .withSlice() references.
-    // The tagging loop should skip acts from different files.
-    const files: FileTab[] = [
-      {
-        path: "src/states.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({}))
-  .emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E")
-  .build();
-`,
-      },
-      {
-        path: "src/slices.ts",
-        content: `
-import { slice } from "@rotorsoft/act";
-import { S } from "./states.js";
-export const MySlice = slice().withState(S).build();
-`,
-      },
-      {
-        path: "src/act1.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { MySlice } from "./slices.js";
-export const app1 = act().withSlice(MySlice).build();
-`,
-      },
-      {
-        path: "src/act2.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { MySlice } from "./slices.js";
-export const app2 = act().withSlice(MySlice).build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    // Both acts should have built
-    expect(model.entries.length).toBeGreaterThanOrEqual(1);
-  });
-
-  it("buildState with null events and actions (lines 28, 33 ?? fallback)", () => {
-    const stateObj = {
-      _tag: "State",
-      name: "NullState",
-      events: null, // triggers ?? {} fallback at line 28
-      actions: null, // triggers ?? {} fallback at line 33
-      given: null,
-      patches: null,
-    };
-    const result = {
-      states: [stateObj],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const st = model.states.find((s) => s.name === "NullState");
-    expect(st).toBeDefined();
-    expect(st!.events).toHaveLength(0);
-    expect(st!.actions).toHaveLength(0);
-  });
-
-  it("buildState with events but null patches (line 47 ?? false fallback)", () => {
-    const stateObj = {
-      _tag: "State",
-      name: "NoPatchState",
-      events: { Evt1: {} },
-      actions: {},
-      given: {},
-      patches: null, // triggers ?. ?? false at line 47
-    };
-    const result = {
-      states: [stateObj],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const st = model.states.find((s) => s.name === "NoPatchState");
-    expect(st).toBeDefined();
-    expect(st!.events[0].hasCustomPatch).toBe(false);
-  });
-
-  it("fixupReactions: handler name in source but no matching fallback (line 95 false)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "TestSlice",
-          states: [],
-          projections: [],
-          reactions: [
-            {
-              event: "SomeOtherEvent",
-              handlerName: "on SomeOtherEvent",
-              dispatches: [],
-              isVoid: false,
-            },
-          ],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const files: FileTab[] = [
-      {
-        path: "src/s.ts",
-        // Source has .on("Created").do(onCreated) but reaction is for SomeOtherEvent
-        content: `.on("Created").do(onCreated).to("items")`,
-      },
-    ];
-    const { model } = buildModel(result, files, new Map());
-    // Handler name stays as "on SomeOtherEvent" since no match for "Created"
-    expect(model.slices[0].reactions[0].handlerName).toBe("on SomeOtherEvent");
-  });
-
-  it("step 1 catch with non-Error thrown (line 145 false branch)", () => {
-    const corruptState = {
-      _tag: "State",
-      get name(): string {
-        // eslint-disable-next-line @typescript-eslint/only-throw-error
-        throw { toString: () => "string error" };
-      },
-      events: {},
-      actions: {},
-    };
-    const result = {
-      states: [corruptState],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model).toBeDefined();
-  });
-
-  it("act builder with undefined states array (line 151 ?? fallback)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: undefined, // triggers ?? [] at line 151
-          slices: [],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model).toBeDefined();
-  });
-
-  it("act builder catch with non-Error (line 157 false branch)", () => {
-    const corruptState = {
-      _tag: "State",
-      get name(): string {
-        // eslint-disable-next-line @typescript-eslint/only-throw-error
-        throw { toString: () => "corrupt string" };
-      },
-      events: {},
-      actions: {},
-    };
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [corruptState],
-          slices: [],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model).toBeDefined();
-  });
-
-  it("slice with undefined _varName uses 'slice' default (line 167)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          // _varName undefined → defaults to "slice" at line 167
-          states: [],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model.slices[0].name).toBe("slice");
-  });
-
-  it("slice with null state and fileError (line 180)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "BrokenSlice",
-          states: [null],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map([["src/broken.ts", "Import failed"]]),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "BrokenSlice");
-    expect(sl!.error).toContain("Import failed");
-  });
-
-  it("slice state buildState throws non-Error (line 205 false branch)", () => {
-    const corruptState = {
-      _tag: "State", // must have _tag so it passes typeof check
-      get name(): string {
-        // eslint-disable-next-line @typescript-eslint/only-throw-error
-        throw { toString: () => "string thrown in buildState" };
-      },
-      events: {},
-      actions: {},
-    };
-    const result = {
-      states: [], // not pre-built
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "SliceWithCorruptState",
-          states: [corruptState],
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "SliceWithCorruptState");
-    expect(sl!.error).toContain("string thrown in buildState");
-  });
-
-  it("slice with undefined states array (line 180 ?? fallback)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "NullStatesSlice",
-          states: undefined as any, // triggers ?? [] at line 180
-          projections: [],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "NullStatesSlice");
-    expect(sl).toBeDefined();
-    expect(sl!.states).toHaveLength(0);
-  });
-
-  it("slice catch with non-Error (line 205 false branch)", () => {
-    const throwingSlice = {
-      _tag: "Slice",
-      _varName: "ThrowSlice2",
-      get states(): any {
-        // eslint-disable-next-line @typescript-eslint/only-throw-error
-        throw { toString: () => "string thrown" };
-      },
-      projections: [],
-      reactions: [],
-    };
-    const result = {
-      states: [],
-      slices: [throwingSlice],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "ThrowSlice2");
-    expect(sl!.error).toBe("string thrown");
-  });
-
-  it("slice with undefined projections (line 213 ?? fallback)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "NoProjSlice",
-          states: [],
-          projections: undefined as any,
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "NoProjSlice");
-    expect(sl!.projections).toHaveLength(0);
-  });
-
-  it("slice projection with non-Projection _tag (line 213-214)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "Sl",
-          states: [],
-          projections: [{ _tag: "NotAProjection", target: "fake" }, null],
-          reactions: [],
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "Sl");
-    expect(sl!.projections).toHaveLength(0);
-  });
-
-  it("slice with undefined reactions (line 222)", () => {
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "Sl",
-          states: [],
-          projections: [],
-          reactions: undefined as any,
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    const sl = model.slices.find((s) => s.name === "Sl");
-    expect(sl!.reactions).toEqual([]);
-  });
-
-  it("slice catch with non-Error from fixupReactions (line 227)", () => {
-    // fixupReactions at the top of the try block — if it throws,
-    // the inner try/catch at line 172-174 catches it
-    // But we need the OUTER catch at line 226 to fire with non-Error
-    // Actually line 227 is about s._sourceFile ?? undefined
-    // Let me check...
-    const result = {
-      states: [],
-      slices: [
-        {
-          _tag: "Slice",
-          _varName: "Sl",
-          states: [],
-          projections: [],
-          reactions: [],
-          // _sourceFile undefined → at line 227 'file: s._sourceFile as string | undefined'
-        },
-      ],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model.slices[0].file).toBeUndefined();
-  });
-
-  it("expectedSlices produces error from result.error or fallback (lines 250-252)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: "Global error",
-      fileErrors: new Map([["src/sl.ts", "file-specific error"]]),
-    };
-    const expectedSlices = new Map([
-      ["MissingSlice", "src/sl.ts"],
-      ["AnotherMissing", "src/other.ts"],
-    ]);
-    const { model } = buildModel(result, [], expectedSlices);
-    const sl1 = model.slices.find((s) => s.name === "MissingSlice");
-    expect(sl1!.error).toBe("file-specific error");
-    const sl2 = model.slices.find((s) => s.name === "AnotherMissing");
-    // Falls through to result.error since no file error for src/other.ts
-    expect(sl2!.error).toBe("Global error");
-  });
-
-  it("expectedSlices fallback to 'Failed to build slice' (line 252)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const expectedSlices = new Map([["Missing", "src/m.ts"]]);
-    const { model } = buildModel(result, [], expectedSlices);
-    const sl = model.slices.find((s) => s.name === "Missing");
-    expect(sl!.error).toBe("Failed to build slice");
-  });
-
-  it("act with undefined reactions (line 290 ?? fallback)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [],
-          slices: [],
-          projections: [],
-          reactions: undefined as any,
-          _sourceFile: "src/app.ts",
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model.reactions).toHaveLength(0);
-  });
-
-  it("act entry with reactions fallback (line 300 ?? fallback)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [],
-          slices: [],
-          projections: [],
-          reactions: undefined as any,
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model.entries).toHaveLength(1);
-    expect(model.entries[0].reactions).toEqual([]);
-  });
-
-  it("scoped import falls all the way to unknownModuleProxy (line 107 branch #6)", () => {
-    // All fileExports.get patterns must fail, reaching the final unknownModuleProxy
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { thing } from "@scope/doesnotexist";
-const x = thing;
-export const S = state({ S: z.object({}) })
-  .init(() => ({})).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model, error } = extractModel(files);
-    expect(error).toBeUndefined();
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("scoped import with subpath (line 103 truthy branch)", () => {
-    const files: FileTab[] = [
-      {
-        path: "packages/mylib/src/utils.ts",
-        content: `export const helper = "val";`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import { helper } from "@myorg/mylib/src/utils";
-const x = helper;
-export const S = state({ S: z.object({ x: z.string() }) })
-  .init(() => ({ x })).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model.states).toHaveLength(1);
-  });
-
-  it("scoped import with no package name after @ (line 102 false)", () => {
-    // Import like "@justscope" with no / after it
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import something from "@nopkg";
-const x = something;
-export const S = state({ S: z.object({}) })
-  .init(() => ({ x })).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("scoped import with empty package name (line 102 false)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/app.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-import something from "@/utils";
-const x = something;
-export const S = state({ S: z.object({}) })
-  .init(() => ({ x })).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-    ];
-    const { error } = extractModel(files);
-    expect(error).toBeUndefined();
-  });
-
-  it("duplicate .withSlice() in same file (line 149 false branch)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/states.ts",
-        content: `
-import { state } from "@rotorsoft/act";
-import { z } from "zod";
-export const S = state({ S: z.object({}) })
-  .init(() => ({})).emits({ E: z.object({}) })
-  .on({ a: z.object({}) }).emit("E").build();
-`,
-      },
-      {
-        path: "src/slices.ts",
-        content: `
-import { slice } from "@rotorsoft/act";
-import { S } from "./states.js";
-export const MySlice = slice().withState(S).build();
-`,
-      },
-      {
-        path: "src/app.ts",
-        content: `
-import { act } from "@rotorsoft/act";
-import { MySlice } from "./slices.js";
-export const app = act().withSlice(MySlice).withSlice(MySlice).build();
-`,
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model).toBeDefined();
-  });
-
-  it("eval throws non-Error (line 182 false branch)", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/bad.ts",
-        content: [
-          'import { state } from "@rotorsoft/act";',
-          'import { z } from "zod";',
-
-          'throw "non-error string";',
-          "export const S = state({ S: z.object({}) })",
-          "  .init(() => ({})).emits({ E: z.object({}) })",
-          '  .on({ a: z.object({}) }).emit("E").build();',
-        ].join("\n"),
-      },
-    ];
-    const { model } = extractModel(files);
-    // File error should be recorded
-    expect(model).toBeDefined();
-  });
-
-  it("act with null slices array (line 190 || fallback)", () => {
-    const result = {
-      states: [],
-      slices: [],
-      projections: [],
-      acts: [
-        {
-          _tag: "Act",
-          states: [],
-          slices: null as any, // triggers || [] at line 190
-          projections: [],
-          reactions: [],
-          _sourceFile: "src/app.ts",
-        },
-      ],
-      error: undefined,
-      fileErrors: new Map<string, string>(),
-    };
-    const { model } = buildModel(result, [], new Map());
-    expect(model).toBeDefined();
-  });
-
-  it("outer catch with non-Error (line 209 false branch)", () => {
-    const realFiles = [{ path: "src/app.ts", content: `const x = 1;` }];
-    const poisoned: any = [...realFiles];
-    poisoned.filter = () => {
-      throw new Error("string error in setup");
-    };
-    const { error } = extractModel(poisoned as FileTab[]);
-    expect(error).toBe("string error in setup");
-  });
-
-  it("outer catch in execute() for setup failure (line 209)", () => {
-    const realFiles = [{ path: "src/app.ts", content: `const x = 1;` }];
-    const poisoned: any = [...realFiles];
-    poisoned.filter = () => {
-      throw new Error("setup failure");
-    };
-    const { error } = extractModel(poisoned as FileTab[]);
-    expect(error).toBeDefined();
-  });
-
-  it("outer catch non-Error branch (line 209 String(e))", () => {
-    const poisoned: any = [{ path: "src/app.ts", content: "const x = 1;" }];
-    poisoned.filter = () => {
-      // eslint-disable-next-line @typescript-eslint/only-throw-error
-      throw { toString: () => "non-error setup failure" };
-    };
-    const { error } = extractModel(poisoned as FileTab[]);
-    expect(error).toBeDefined();
-  });
-
-  it("eval catch non-Error branch (line 182 String(e))", () => {
-    const files: FileTab[] = [
-      {
-        path: "src/bad.ts",
-        content: [
-          'import { state } from "@rotorsoft/act";',
-          'import { z } from "zod";',
-          'throw "non-error throw";',
-          "export const S = state({ S: z.object({}) })",
-          "  .init(() => ({})).emits({ E: z.object({}) })",
-          '  .on({ a: z.object({}) }).emit("E").build();',
-        ].join("\n"),
-      },
-    ];
-    const { model } = extractModel(files);
-    expect(model).toBeDefined();
-  });
-});
